requirements.txt
opencv-python
langchain
langchain-openai
langchain-community
python-dotenv
pyaudio
soundfile
SpeechRecognition
git+https://github.com/openai/whisper.git
openai
loguru
